[SCRUM-30] feat: Criando visualização de site para usuários não logados e pág. inicial com guia educativo
