+++ .gitignore
@@ -22,3 +22,7 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+#SonarQube
+sonar-project.properties
+.scannerwork
