+++ src/components/Footer/Footer.tsx
@@ -1,10 +1,12 @@
 import "./styles.scss";
 import { IconBrandGithubFilled } from "@tabler/icons-react";
 
-export default function Footer() {
+export default function Footer({ autenticado = true }: { autenticado?: boolean }) {
+    
+    const footerClass = autenticado ? "footer_container" : "footer_container sem_login";
 
     return (
-        <footer className="footer_container">
+        <footer className={footerClass}>
             <div className="footer_cima">
                 <div className="footer_esq">
                     <h4>API 2025.01</h4>
@@ -12,9 +14,12 @@ export default function Footer() {
                 </div>
                 <div className="footer_dir">
                     <img src="../../logo_hive.svg" />
-                    <a className="footer_botao" href="https://github.com/TeamHiveAPI/API-2025.01" target="_blank">
-                    <IconBrandGithubFilled color="white"/> 
-                    GitHub </a>
+                    <a className="footer_botao" href="https://github.com/TeamHiveAPI/API-2025.01" 
+                    target="_blank" 
+                    rel="noopener noreferrer">
+                        <IconBrandGithubFilled color="white" />
+                        GitHub
+                    </a>
                 </div>
             </div>
             <div className="footer_baixo">
@@ -22,5 +27,5 @@ export default function Footer() {
                 <p>Política de Privacidade</p>
             </div>
         </footer>
-    )
+    );
 }
