+++ src/components/Footer/styles.scss
@@ -11,6 +11,12 @@
     background-color: #F2F2F2;
 }
 
+.sem_login {
+    margin-left: 0px !important;
+    width: 100% !important;
+    margin-top: 96px;
+}
+
 .footer_cima {
     display: flex;
     justify-content: space-between;
