+++ src/components/Navbar/styles.scss
@@ -1,0 +1,40 @@
+@use "../../global.scss" as *;
+
+.nav_wrapper {
+    display: flex;
+    justify-content: center;
+    position: fixed;
+    top: 0;
+    z-index: 1000;
+    display: flex;
+    width: 100%;
+    padding: 12px 0px;
+    background-color: #FFF;
+    box-shadow: 0px 1px 4px 2px #1A1A1A10;
+}
+
+.nav_container {
+    display: flex;
+    width: 1220px;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.nav_meio {
+    display: flex;
+    gap: 32px;
+}
+
+.nav_meio a {
+    color: #606060;
+    font-family: "Poppins";
+}
+
+.nav_meio a:hover {
+    color: $roxo-hover;
+}
+
+.nav_ativo {
+    color: $roxo !important;
+    font-weight: 500;
+}
