+++ src/pages/Aprendizagem/styles.scss
@@ -1,0 +1,51 @@
+@use "../../global.scss" as *;
+
+.banner {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    width: 100%;
+    height: calc(100dvh - 72px);
+    overflow: hidden;
+    margin-top: 72px;
+  }
+  
+  .video_fundo {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    z-index: -1;
+  }
+  
+  .conteudo_banner {
+    width: 1220px;
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    z-index: 1;
+    color: white;
+  }
+
+.conteudo_banner h1 {
+    font-size: 56px;
+    font-weight: 600;
+    font-family: "Poppins";
+    width: 70%;
+}
+
+.conteudo_banner img {
+    margin-top: 48px;
+    margin-bottom: 48px;
+    width: 40%;
+}
+
+.conteudo_banner p {
+    font-size: 18px;
+    font-family: "Poppins";
+    margin-bottom: 48px;
+}
+
