+++ src/pages/Dashboard/styles.scss
@@ -1,0 +1,158 @@
+@use "../../global.scss" as *;
+
+.home_cima {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    margin-bottom: 20px;
+    gap: 32px;
+}
+
+.home_cima_alerta {
+    background-color: #FAFAFA;
+    padding: 12px;
+    border-radius: 4px;
+    font-size: 18px;
+    font-weight: 600;
+    color: #404040;
+    font-family: Poppins;
+}
+
+.home_cima button {
+    box-shadow: 0px 4px 0px 0px #3a3599;
+    min-width: 15%;
+}
+
+.home_cima button:hover {
+    box-shadow: 0px 4px 0px 0px #292576;
+}
+
+.mensagem img {
+    width: 100%;
+}
+
+.caes_wrapper.caal.mini {
+    padding: 12px;
+    margin-bottom: 0px;
+    width: 100%;
+    border-radius: 8px;
+}
+
+.caes_wrapper.caal.mini:hover {
+    filter: none;
+    cursor: auto;
+}
+
+.caal_titulo.mini {
+    font-size: 16px;
+}
+
+.caal_estacao_mini {
+    font-size: 14px !important;
+}
+
+.home_cima_titulo h2 {
+    color: #808080;
+    font-size: 16px;
+    font-weight: 500;
+    margin-bottom: 20px;
+}
+
+.dashboard_cima {
+    display: flex;
+    gap: 32px;
+    margin-top: 32px;
+}
+
+.dashboard_card {
+    display: flex;
+    flex-direction: column;
+    border-radius: 16px;
+    border: 1px solid #F5F5F5;
+    background: #FCFCFC;
+    box-shadow: 0px 4px 0px 0px #E8E8E8;
+    width: 25%;
+    padding: 20px;
+}
+
+.dashboard_card_cima {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+}
+
+.dashboard_card_cima p {
+    font-family: "Poppins";
+    font-weight: 500;
+    font-size: 20px;
+    color: #606060;
+}
+
+.dashboard_numero {
+    margin-top: 12px;
+    margin-bottom: -6px;
+    font-family: Poppins;
+    font-weight: 600;
+    font-size: 36px;
+    color: $roxo;
+}
+
+.dashboard_grafico {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+    margin-top: 32px;
+}
+
+.dashboard_grafico_baixo {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 32px;
+}
+
+.dashboard_grafico_unidade {
+    width: 100%;
+    border-radius: 16px;
+    border: 1px solid #F5F5F5;
+    background: #FCFCFC;
+    box-shadow: 0px 4px 0px 0px #E8E8E8;
+    padding: 20px;
+    transition: 0.3s ease;
+}
+
+.dashboard_grafico_cima {
+    display: flex;
+    justify-content: flex-start;
+    border-radius: 16px;
+    border: 1px solid #F5F5F5;
+    background: #FCFCFC;
+    box-shadow: 0px 4px 0px 0px #E8E8E8;
+    padding: 20px;
+}
+
+.erro_datas {
+    font-size: 16px;
+    line-height: 28px;
+    font-weight: 500;
+    color: #606060;
+    text-align: center;
+}
+
+.home_alerta_nenhum {
+    padding: 12px;
+    background-color: #FCFCFC;
+    width: 100%;
+    box-shadow: 0px 4px 0px 0px #E8E8E8;
+    border-radius: 8px;
+    font-family: "Poppins";
+    color: #606060;
+}
+
+.dashboard_grafico_wrapper {
+    display: flex;
+    flex-direction: column;
+}
+
+.gap_32 {
+    margin-bottom: 32px;
+}
